Update to Main.java: Brute Force line 162
I believe this proposed file change effectively solves our lingering issue in which the output hull for Brute Force contained extra elements.  This change consists simply of adding the condition that an evaluated point may not be ON the line; it must be to one side or the other.  Nima, thanks for your flash of brilliance when you pointed out earlier that the offending point seemed to be very close to the hull.